fix: 尝试解决deploy无onclick的问题
未解决
